pyproject.toml
[project]
name = "reddit-actions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "praw>=7.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]


[tool.ruff]
line-length = 120
exclude = [".git", ".venv"]

[tool.ruff.lint]
select = [
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "C90", # mccabe (we are not ready)
    "D", # pydocstyle
    "E", # pycodestyle
    "F", # pyflakes
    "FBT", # flake8-boolean-trap
    "I", # isort
    "N", # pep8-naming
    "PIE", # flake8-pie
    "PL", # pylint
    "SIM", # flake8-simplify
    "UP", # pyupgrade
    "YTT", # flake8-2020
]
ignore = [
    "D203", # One blank line before class
    "D212", # Multiline summary second line
    "FBT001", # Boolean-typed positional argument in function definition
    "FBT002", # Boolean default positional argument in function definition
]

[tool.ruff.lint.isort]
known-first-party = ["pytest_fixtures_fixtures"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "B011", # Do not `assert False`
    "D401", # First line of docstring should be in imperative mood
    "D404", # First word of the docstring should not be "This"
    "F841", # Local variable is assigned to but never used
    "PLR2004", # Magic values
    "PLR0913", # Too many arguments in function definition
    "PLC0415", # `import` should be at the top-level of a file
    "SIM117", # Use a single `with` statement
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
    "D203", # One blank line before class
    "D212", # Multiline summary second line
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
